moved edge_buffed_mask
